Fix parameter logic and introduce a "namespace"
Parameter logic being fixed is critical, or at least requires you to understand that you have to work around it by passing `-Localhost:$False` to the function when you call it with both the -ComputerName and the -PSRemoting flags. Silly error.

"Namespace" being "ST" for Svendsen Tech. Get-STDotNetVersion is now the primary function name, and if you have another Get-DotNetVersion command in your session, that won't be overwritten - but to preserve backwards compatibility, I am aliasing Get-DotNetVersion to the old command ''if it does not exist''...