Unbreak PSRemoting feature
My sincere apologies. Good I didn't update in the PSGallery yet. The PSRemoting feature was broken as the helper function was not in the scope of the remote computer. I did not test until now. I added the helper function to the SetDataHashObject function and it now works.

Nobody gave me feedback, so hopefully not too many were affected - and used the PSGallery to install. This became messy. Sorry again. I hope this works. Please give me feedback. :(